services/generate-description/pyproject.toml
[project]
name = "gerenate-product-description"
version = "0.1.0"
description = "description generator"
authors = [
    {name = "elias torres",email = "elias.qlr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (==0.104.1)",
    "uvicorn[standard] (==0.24.0)",
    "python-dotenv (==1.0.0)",
    "httpx (==0.25.2)",
    "transformers (==4.54.1)",
    "accelerate (==1.9.0)",
    "torch (==2.7.1+cu128)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
